+++ docs/_client/src/components/menubar/index.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-unused-vars */
 import getClassnames from 'js/getClassnames';
+import logEventDetail from 'js/logEventDetail';
 import { MenuBar } from 'root';
 import './menubar.scss';
 
@@ -8,6 +9,11 @@ const { list } = getClassnames(siteClassNames.menubar);
 
 // Get the elements.
 const menuBarList = document.querySelector(list);
+
+// Report event details.
+menuBarList.addEventListener('init', logEventDetail);
+menuBarList.addEventListener('stateChange', logEventDetail);
+menuBarList.addEventListener('destroy', logEventDetail);
 
 // Create the MenuBar.
 const menuBar = new MenuBar(menuBarList);