+++ docs/_client/src/components/menubutton/index.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-unused-vars */
 import getClassnames from 'js/getClassnames';
+import logEventDetail from 'js/logEventDetail';
 import { MenuButton } from 'root';
 import './menubutton.scss';
 
@@ -8,6 +9,11 @@ const { button } = getClassnames(siteClassNames.menubutton);
 
 // Get the elements.
 const controller = document.querySelector(button);
+
+// Report event details.
+controller.addEventListener('init', logEventDetail);
+controller.addEventListener('stateChange', logEventDetail);
+controller.addEventListener('destroy', logEventDetail);
 
 // Create the MenuButton.
 const menuButton = new MenuButton(controller);