+++ src/Disclosure/index.js
@@ -50,27 +50,6 @@ export default class Disclosure extends AriaComponent {
        * @type {boolean}
        */
       allowOutsideClick: true,
-
-      /**
-       * Callback to run after the component initializes.
-       *
-       * @callback initCallback
-       */
-      onInit: () => {},
-
-      /**
-       * Callback to run after component state is updated.
-       *
-       * @callback stateChangeCallback
-       */
-      onStateChange: () => {},
-
-      /**
-       * Callback to run after the component is destroyed.
-       *
-       * @callback destroyCallback
-       */
-      onDestroy: () => {},
     };
 
     // Merge remaining options with defaults and save all as instance properties.
@@ -170,8 +149,8 @@ export default class Disclosure extends AriaComponent {
      */
     tabIndexDeny(this.interactiveChildElements);
 
-    // Run {initCallback}
-    this.onInit.call(this);
+    // Fire the init event.
+    this.dispatchEventInit();
   }
 
   /**
@@ -298,8 +277,8 @@ export default class Disclosure extends AriaComponent {
     // Reset initial state.
     this.state = { expanded: this.loadOpen };
 
-    // Run {destroyCallback}
-    this.onDestroy.call(this);
+    // Fire the destroy event.
+    this.dispatchEventDestroy();
   }
 
   /**