+++ src/Listbox/index.js
@@ -31,36 +31,8 @@ export default class ListBox extends Popup {
     this.controller = controller;
     this.target = super.constructor.getTargetElement(controller);
 
-    /**
-     * Options shape.
-     *
-     * @type {object}
-     */
-    const defaultOptions = {
-      /**
-       * Callback to run after the component initializes.
-       *
-       * @callback initCallback
-       */
-      onInit: () => {},
-
-      /**
-       * Callback to run after component state is updated.
-       *
-       * @callback stateChangeCallback
-       */
-      onStateChange: () => {},
-
-      /**
-       * Callback to run after the component is destroyed.
-       *
-       * @callback destroyCallback
-       */
-      onDestroy: () => {},
-    };
-
-    // Merge remaining options with defaults and save all as instance properties.
-    Object.assign(this, defaultOptions, options);
+    // Merge options as instance properties.
+    Object.assign(this, options);
 
     // Bind class methods.
     this.preventWindowScroll = this.preventWindowScroll.bind(this);
@@ -145,8 +117,8 @@ export default class ListBox extends Popup {
     // Prevent scrolling when using UP/DOWN arrows on the button
     window.addEventListener('keydown', this.preventWindowScroll);
 
-    // Run {initCallback}
-    this.onInit.call(this);
+    // Fire the init event.
+    this.dispatchEventInit();
   }
 
   /**
@@ -409,7 +381,7 @@ export default class ListBox extends Popup {
     this.target.removeEventListener('blur', this.targetHandleBlur);
     window.removeEventListener('keydown', this.preventWindowScroll);
 
-    // Run {destroyCallback}
-    this.onDestroy.call(this);
+    // Fire the destroy event.
+    this.dispatchEventDestroy();
   }
 }