+++ CHANGELOG.md
@@ -16,16 +16,20 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Gets Dialog content element(s) if none provided (#51)
 - Uses `[Symbol.toStringTag]` for component identification via `instance.toString()` (#52)
 - Exports `isInstanceOf` helper function (#52)
+- Component element fires `init`, `stateChange` and `destroy` events (#55)
 
 **Fixed**
 
 - MenuButton could attempt to focus the first Menu child even when the Popup is closed (14599f0)
 - Component callbacks could be run more than once (#54)
+- MenuBar wasn't tracking the current Popup (6c2fe90)
+- MenuBar keydown left/right could trigger a Popup `stateChange` event (e86d06e)
 
 **Removed**
 
 - Dialog no longer requires a close button, and will not create one (#51)
-- Popup no longer accepts `onInit` and `onDestroy` callbacks; additionally, MenuBar no longer acepts `onPopupInit` (#54)
+- Components no longer accept `onInit`, `onStateChange`, nor `onDestroy` callbacks; additionally, MenuBar no longer acepts `onPopupInit` (#54, #55)
+- MenuBar no longer tracks Popup `expanded` state separately (e86d06e)
 
 ## 0.3.2